requirements.txt
streamlit>=1.32
pillow>=10.0